UNLOCK ENCRYPT AES AND ADD DATA TEST

+++ BE/aquariumshop_2/src/main/kotlin/com/aquariumshop/aquariumshop/service/impl/ProductServiceImpl.kt
@@ -70,7 +70,7 @@ class ProductServiceImpl(
             commentOfProduct.forEach { item ->
                 ratingSum += item.rating!!
             }
-            product.rating = ratingSum.toFloat() / commentOfProduct.size
+            product.rating = ratingSum.toFloat() / commentOfProduct.size.toFloat()
             productRepository.save(product)
         }
